pyproject.toml
[project]
name = "emoji-trigger-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.8",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
